src/requirements.txt
fastapi==0.110.0
pymongo[srv]==4.6.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
uvicorn==0.27.0